model ka change h iskay bd

+++ app/src/main/java/com/teamx/hatlyDriver/data/dataclasses/totalEarning/TotalEarningsData.kt
@@ -2,11 +2,11 @@ package com.teamx.hatlyDriver.data.dataclasses.totalEarning
 
 data class TotalEarningsData(
     val _id: String,
-    val totalEarning: Int,
-    val totalOrders: Int,
-    val totalOrdersEarning: Int,
-    val totalParcels: Int,
-    val totalParcelsEarning: Int,
+    val totalEarning: Double,
+    val totalOrders: Double,
+    val totalOrdersEarning: Double,
+    val totalParcels: Double,
+    val totalParcelsEarning: Double,
     val totalSpendTime: TotalSpendTime,
     val userId: UserId
 )
